not removing READ for anyone that has EDITted the request

+++ arc-expenses-backend/service/src/main/java/arc/expenses/service/AclService.java
@@ -32,7 +32,7 @@ public class AclService extends JdbcMutableAclService{
             MutableAcl acl = (MutableAcl) readAclById(objectIdentity);
             for(int i=0; i<acl.getEntries().size();i++){
                 for(Sid sid : principals) {
-                    if(acl.getEntries().get(i).getSid().equals(sid)){
+                    if(acl.getEntries().get(i).getSid().equals(sid) && !acl.getEntries().get(i).getPermission().equals(ArcPermission.READ)){
                         acl.deleteAce(i);
                         i--;
                     }
